FSL FNIRT

FSL FNIRT wrapper for non-linear registration

Usage

ni_fsl_fnirt(
  in_file,
  ref_file,
  affine_file = NULL,
  apply_inmask = NULL,
  apply_intensity_mapping = NULL,
  apply_refmask = NULL,
  args = NULL,
  bias_regularization_lambda = NULL,
  biasfield_resolution = NULL,
  config_file = NULL,
  derive_from_ref = NULL,
  field_file = NULL,
  fieldcoeff_file = NULL,
  hessian_precision = NULL,
  in_fwhm = NULL,
  in_intensitymap_file = NULL,
  inmask_file = NULL,
  inmask_val = NULL,
  intensity_mapping_model = NULL,
  intensity_mapping_order = NULL,
  inwarp_file = NULL,
  jacobian_file = NULL,
  jacobian_range = NULL,
  log_file = NULL,
  max_nonlin_iter = NULL,
  modulatedref_file = NULL,
  out_intensitymap_file = NULL,
  ref_fwhm = NULL,
  refmask_file = NULL,
  refmask_val = NULL,
  regularization_lambda = NULL,
  regularization_model = NULL,
  skip_implicit_in_masking = NULL,
  skip_implicit_ref_masking = NULL,
  skip_inmask = NULL,
  skip_intensity_mapping = NULL,
  skip_lambda_ssq = NULL,
  skip_refmask = NULL,
  spline_order = NULL,
  subsampling_scheme = NULL,
  warp_resolution = NULL,
  warped_file = NULL,
  .cwd = NULL,
  .env = NULL,
  .engine = NULL,
  .profile = NULL,
  dry_run = FALSE,
  echo = interactive()
)

Arguments

in_file

Character; file path. name of input image Required.

ref_file

Character; file path. name of reference image Required.

affine_file

Character; file path. name of file containing affine transform

apply_inmask

Character or numeric vector. list of iterations to use input mask on (1 to use, 0 to skip)

apply_intensity_mapping

Character or numeric vector. List of subsampling levels to apply intensity mapping for (0 to skip, 1 to apply)

apply_refmask

Character or numeric vector. list of iterations to use reference mask on (1 to use, 0 to skip)

args

Character. Additional parameters to the command

bias_regularization_lambda

Numeric. Weight of regularisation for bias-field, default 10000

biasfield_resolution

Character or numeric vector. Resolution (in mm) of bias-field modelling local intensities, default 50, 50, 50

config_file

Character or numeric vector. Name of config file specifying command line arguments

derive_from_ref

Logical. If true, ref image is used to calculate derivatives. Default false

field_file

Character or numeric vector. name of output file with field or true

fieldcoeff_file

Character or numeric vector. name of output file with field coefficients or true

hessian_precision

Character; one of: "double", "float". Precision for representing Hessian, double or float. Default double

in_fwhm

Character or numeric vector. FWHM (in mm) of gaussian smoothing kernel for input volume, default [6, 4, 2, 2]

in_intensitymap_file

Character or numeric vector. name of file/files containing initial intensity mapping usually generated by previous fnirt run

inmask_file

Character; file path. name of file with mask in input image space

inmask_val

Numeric. Value to mask out in –in image. Default =0.0

intensity_mapping_model

Character; one of: "none", "global_linear", "global_non_linear", "local_linear", "global_non_linear_with_bias", "local_non_linear". Model for intensity-mapping

intensity_mapping_order

Integer. Order of poynomial for mapping intensities, default 5

inwarp_file

Character; file path. name of file containing initial non-linear warps

jacobian_file

Character or numeric vector. name of file for writing out the Jacobian of the field (for diagnostic or VBM purposes)

jacobian_range

Character or numeric vector. Allowed range of Jacobian determinants, default 0.01, 100.0

log_file

Character; file path. Name of log-file

max_nonlin_iter

Character or numeric vector. Max # of non-linear iterations list, default [5, 5, 5, 5]

modulatedref_file

Character or numeric vector. name of file for writing out intensity modulated –ref (for diagnostic purposes)

out_intensitymap_file

Character or numeric vector. name of files for writing information pertaining to intensity mapping

ref_fwhm

Character or numeric vector. FWHM (in mm) of gaussian smoothing kernel for ref volume, default [4, 2, 0, 0]

refmask_file

Character; file path. name of file with mask in reference space

refmask_val

Numeric. Value to mask out in –ref image. Default =0.0

regularization_lambda

Character or numeric vector. Weight of regularisation, default depending on –ssqlambda and –regmod switches. See user documentation.

regularization_model

Character; one of: "membrane_energy", "bending_energy". Model for regularisation of warp-field [membrane_energy bending_energy], default bending_energy

skip_implicit_in_masking

Logical. skip implicit masking based on value in –in image. Default = 0

skip_implicit_ref_masking

Logical. skip implicit masking based on value in –ref image. Default = 0

skip_inmask

Logical. skip specified inmask if set, default false

skip_intensity_mapping

Logical. Skip estimate intensity-mapping default false

skip_lambda_ssq

Logical. If true, lambda is not weighted by current ssq, default false

skip_refmask

Logical. Skip specified refmask if set, default false

spline_order

Integer. Order of spline, 2->Qadratic spline, 3->Cubic spline. Default=3

subsampling_scheme

Character or numeric vector. sub-sampling scheme, list, default [4, 2, 1, 1]

warp_resolution

Character or numeric vector. (approximate) resolution (in mm) of warp basis in x-, y- and z-direction, default 10, 10, 10

warped_file

Character; file path. name of output image

.cwd

Working directory override.

.env

Named character vector of environment variables.

.engine

Execution engine override.

.profile

Runtime profile override.

dry_run

Logical; preview command without executing.

echo

Logical; echo stdout/stderr in real time.

Value

An ni_result object.